* restructured floor initialization, renderer handling, and general cleanup:    * initialization now requires a init_state struct which contains various information that can or must be set (has sane defaults, call and data path have to be set)    * since vulkan rendering (swapchain) can't be used in conjunction with opengl rendering (swapchain), a renderer has to be chosen now on init -> this will generally default to opengl unless vulkan has been explicitly set or the compute backend set in the config is vulkan (+the renderer is set to default)    * if the vulkan renderer is used, this will create a vulkan_compute context with renderer support, i.e. will set up a surface and swapchain based on the sdl video driver and current window    * if the vulkan renderer is used and the compute backend is also vulkan, both will use the same context (get_compute_context() will return the same as get_vulkan_context())    * can now specify an app name and app version on init (will be passed through to vulkan + will set the app name as the window title/caption directly on window creation)    * can now specify the min vulkan api version that should be used (defaults to 1.0.5)    * init will now return a bool (false if it failed, true if it succeeded)    * renamed start_draw -> start_frame, stop_draw -> end_frame, resize_window -> resize_gl_window    * misc uint64_t -> uint32_t changes where uint32_t is enough    * fixed dpi computation w/o an opengl context    * NOTE: vulkan isn't hooked up to start_frame/end_frame/swap, and I'm not quite sure yet if I should do that (but those functions can be called w/o an opengl context now)  * vulkan_compute: added proper rendering support (swapchain handling, drawable image acquire + present), will only enable/use surface/swapchain extensions now when rendering is enabled  * vulkan_buffer: reenable VK_BUFFER_USAGE_INDEX_BUFFER_BIT usage again (we do of course still need index buffers)  * opencl_compute: platform_index is a uint32_t now + disable opengl sharing when the floor renderer isn't opengl  * json: added support for directly retrieving int32_t/uint32_t values (via saturated convert from the resp. int64_t/uint64_t value)  * event_objects: use uint32_t/uint64_t everywher + window_resize_event_base::size is now a uint2 instead of a size2  * darwin_helper: system version query defaults to os x 10.11 and ios 9.0 now  * updated README (added MinGW/MSYS2 support note, SDL2 2.0.4 is required now, Vulkan headers should at least be version 24)

+++ compute/opencl/opencl_compute.hpp
@@ -36,7 +36,7 @@ public:
 	//////////////////////////////////////////
 	// init / context creation
 	
-	opencl_compute(const uint64_t platform_index = ~0ull,
+	opencl_compute(const uint32_t platform_index = ~0u,
 				   const bool gl_sharing = false,
 				   const vector<string> whitelist = {});
